+++ apps/pizza_api/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+
+### Added
+
+- Configurable API pool settings for db connection(s)
+
+### Changed
+
+- Removed add category logic from API (migrated to worker app)
+- Removed process pending logic from API (migrated to worker app)
+
 ## [0.4.0] - 2026-06-21
 
 ### Changed